feat: add customizable application theme (#542)
* Added a drop down menu on the settings page which allows you to switch the app theme between light, dark, and auto (default).

This is my first time using visual studio as well as working on a UWP app. I could see there was a lot of generated code but wasn't sure what the official way to add to the ui was so I just made changes manually.

It would be nice if the feature could be copied into the actual app (just implemented better then how I did it).

* undo unrelated changes

* refactor

* reorder setting card

* remove whitesapce

---------

Co-authored-by: Tung Huynh <tung75605@gmail.com>